+++ tests/testthat/test-bayesian_as_frequentist.R
@@ -1,8 +1,9 @@
+skip_on_cran()
+skip_if_offline()
+skip_if_not_or_load_if_installed("httr2")
+
 test_that("rstanarm to freq", {
-  skip_on_cran()
-  skip_if_offline()
   skip_if_not_or_load_if_installed("rstanarm")
-  skip_if_not_or_load_if_installed("httr")
 
   set.seed(333)
   m <- insight::download_model("stanreg_glm_1")
@@ -14,11 +15,8 @@ test_that("rstanarm to freq", {
 
 
 test_that("rstanarm to freq", {
-  skip_on_cran()
-  skip_if_offline()
   skip_if_not_or_load_if_installed("rstanarm")
   skip_if_not_or_load_if_installed("lme4")
-  skip_if_not_or_load_if_installed("httr")
 
   set.seed(333)
   m <- insight::download_model("stanreg_lmerMod_1")
@@ -30,12 +28,9 @@ test_that("rstanarm to freq", {
 
 
 test_that("brms beta to freq", {
-  skip_on_cran()
-  skip_if_offline()
   skip_if_not_or_load_if_installed("brms")
   skip_if_not_or_load_if_installed("glmmTMB")
   skip_if_not_or_load_if_installed("lme4")
-  skip_if_not_or_load_if_installed("httr")
   skip_if_not_or_load_if_installed("betareg")
 
   set.seed(333)
@@ -53,13 +48,10 @@ test_that("brms beta to freq", {
 
 
 test_that("ordbetareg to freq", {
-  skip_on_cran()
-  skip_if_offline()
   skip_if_not_or_load_if_installed("brms")
   skip_if_not_or_load_if_installed("ordbetareg")
   skip_if_not_or_load_if_installed("glmmTMB")
   skip_if_not_or_load_if_installed("lme4")
-  skip_if_not_or_load_if_installed("httr")
   skip_if_not_or_load_if_installed("datawizard")
 
   set.seed(333)
@@ -74,4 +66,29 @@ test_that("ordbetareg to freq", {
   m2 <- convert_bayesian_as_frequentist(m)
 
   expect_equal(lme4::fixef(m1), lme4::fixef(m2), tolerance = 1e-1)
+})
+
+
+test_that("brms 0 + Intercept to freq", {
+  skip_if_not_or_load_if_installed("brms")
+
+  set.seed(333)
+  data(mtcars)
+  m <- brms::brm(qsec ~ 0 + Intercept + mpg, data = mtcars, refresh = 0)
+  m1 <- lm(qsec ~ mpg, data = mtcars)
+  m2 <- convert_bayesian_as_frequentist(m)
+
+  expect_equal(coef(m1), coef(m2), tolerance = 1e-2)
+})
+
+
+test_that("brms Interaction terms to freq", {
+  skip_if_not_or_load_if_installed("brms")
+
+  set.seed(333)
+  m <- brms::brm(qsec ~ mpg * as.factor(am), data = mtcars, refresh = 0)
+  m1 <- lm(qsec ~ mpg * as.factor(am), data = mtcars)
+  m2 <- convert_bayesian_as_frequentist(m)
+
+  expect_equal(coef(m1), coef(m2), tolerance = 1e-2)
 })